nice api cleanup
