+++ crates/learner/tests/database/instruction/add.rs
@@ -1,14 +1,4 @@
-use add::Add;
-use chrono::{TimeZone, Utc};
-use learner::{
-  database::*,
-  error::LearnerError,
-  paper::{Author, Paper},
-};
-use query::Query;
-
-use super::setup_test_db;
-use crate::{create_second_test_paper, create_test_paper, traced_test, TestResult};
+use super::*;
 
 /// Basic paper addition tests
 mod basic_operations {
