+++ crates/learner/tests/database/mod.rs
@@ -1,6 +1,3 @@
 use super::*;
 
-mod find;
-mod save;
-
 mod instruction;
